RxNav.jl

Julia interface to the National Library of Medicine's online pharmaceutical RxNav API

Examples

julia> using RxNav

julia> for i in interact("fentanyl", "selegiline")
           if i.severity == "high"
               println(i.description)
           end
       end
Narcotic analgesics - monoamine oxidase (MAO) inhibitors

julia> println(RxNav.getSpellingSuggestions("nortriptelene"))
["nortriptyline", "Nortriptylina"]

julia> interact("1191", "warfarin", "vitamin K")
3-element Vector{NamedTuple}:
 (drug1 = "aspirin", drug2 = "vitamin K", severity = "N/A", description = "Acetylsalicylic acid may decrease the excretion rate of Phylloquinone which could result in a higher serum level.")
 (drug1 = "aspirin", drug2 = "warfarin", severity = "N/A", description = "Acetylsalicylic acid may increase the anticoagulant activities of Warfarin.")
 (drug1 = "vitamin K", drug2 = "warfarin", severity = "N/A", description = "The therapeutic efficacy of Warfarin can be decreased when used in combination with Phylloquinone.")

julia> filter(x -> occursin("Pediatric", x), drugs("riboflavin"))
2-element Vector{String}:
 "alpha-tocopherol acetate 1.4 MG/ML / ascorbic acid 16 MG/ML / biotin 0.004 MG/ML / dexpanthenol 1 MG/ML / ergocalciferol 0.002 MG/ML / folic acid 0.028 MG/ML / niacinamide 3.4 MG/ML / pyridoxine hydrochloride 0.2 MG/ML / riboflavin 0.28 MG/ML / thiamine hydrochloride 0.24 MG/ML / vitamin A 0.14 MG/ML / vitamin B12 0.0002 MG/ML / vitamin K1 0.04 MG/ML Injectable Solution [MVI Pediatric]"
 "alpha-tocopherol acetate 1.4 UNT/ML / ascorbic acid 16 MG/ML / biotin 0.004 MG/ML / cholecalciferol 80 UNT/ML / dexpanthenol 1 MG/ML / folic acid 0.028 MG/ML / niacinamide 3.4 MG/ML / pyridoxine hydrochloride 0.2 MG/ML / riboflavin 0.28 MG/ML / thiamine hydrochloride 0.24 MG/ML / vitamin A palmitate 460 UNT/ML / vitamin B12 0.0002 MG/ML / vitamin K1 0.04 MG/ML Injectable Solution [Infuvite Pediatric]"

Installation

You may install the package from Github in the usual way, or to install the current master copy:

using Pkg
Pkg.add("http://github.com/wherrera10/RxNav.jl")

Functions Reference

See also the National Library of Medicine's RxNorm API documentation. Since RxNav dropped their interactions database in 2023, see also the RxCheck API documentation.

RxNav.RxNav

RxNav.URI_DICT

RxNav.drugs

RxNav.filterByProperty

RxNav.findClassByName

RxNav.findClassesById

RxNav.findRxcuiById

RxNav.findRxcuiByString

RxNav.findSimilarClassesByClass

RxNav.findSimilarClassesByDrugList

RxNav.getAllClasses

RxNav.getAllConcepts

RxNav.getAllConceptsByStatus

RxNav.getAllConceptsByTTY

RxNav.getAllHistoricalNDCs

RxNav.getAllNDCsByStatus

RxNav.getAllProperties

RxNav.getAllRelatedInfo

RxNav.getAllRxTermInfo

RxNav.getApproximateMatch

RxNav.getClassByRxNormDrugId

RxNav.getClassByRxNormDrugName

RxNav.getClassContexts

RxNav.getClassGraphBySource

RxNav.getClassMembers

RxNav.getClassTree

RxNav.getClassTypes

RxNav.getDisplayTerms

RxNav.getDrugs

RxNav.getIdTypes

RxNav.getMultiIngredBrand

RxNav.getNDCProperties

RxNav.getNDCStatus

RxNav.getNDCs

RxNav.getPropCategories

RxNav.getPropNames

RxNav.getProprietaryInformation

RxNav.getRelaTypes

RxNav.getRelas

RxNav.getRelatedByRelationship

RxNav.getRelatedByType

RxNav.getRxConceptProperties

RxNav.getRxNormName

RxNav.getRxNormVersion

RxNav.getRxProperty

RxNav.getRxTermDisplayName

RxNav.getRxTermsVersion

RxNav.getRxcuiHistoryStatus

RxNav.getSimilarityInformation

RxNav.getSourceTypes

RxNav.getSourcesOfDrugClassRelations

RxNav.getSpellingSuggestions

RxNav.getTermTypes

RxNav.getdoc

RxNav.getjson

RxNav.interact

RxNav.interaction

RxNav.interaction_within_list

RxNav.interaction_within_pair

RxNav.is_in_rxcui_format

RxNav.loadAPIkey

RxNav.morearg

RxNav.morearg

RxNav.name

RxNav.rxcui

RxNav.RxNav — Module

RxNav National Library of Medicine REST API wrapper in Julia.

RxNav.URI_DICT — Constant

URI_DICT is a Dict of RxNav REST or RxCheck v1 urls keyed by strings.

RxNav.drugs — Method

drugs(name)

Given a drug name, return a list of all available dosing forms of the drug. Return nothing if the lookup fails.

RxNav.filterByProperty — Function

filterByProperty /rxcui/rxcui/filter Concept RXCUI if the predicate is true Active Return true if the concept RXCUI satisfies the specified property and property values. Return false if it does not fit specified properties, nothing if error occurs.

RxNav.findClassByName — Function

findClassByName

/class/byName Drug classes with a specified class name

RxNav.findClassesById — Method

findClassesById

/class/byId Drug classes with a specified class identifier

RxNav.findRxcuiById — Function

findRxcuiById /rxcui?idtype=...&id=... Concepts associated with a specified identifier Active or Current

RxNav.findRxcuiByString — Function

findRxcuiByString /rxcui?name=... Concepts with a specified name Active or Current

RxNav.findSimilarClassesByClass — Function

findSimilarClassesByClass

/class/similar Classes with similar clinically-significant RxNorm ingredients

RxNav.findSimilarClassesByDrugList — Function

findSimilarClassesByDrugList

/class/similarByRxcuis Classes with clinically-significant RxNorm ingredients similar to a specified list

RxNav.getAllClasses — Function

getAllClasses

/allClasses All classes (may limit by class type)

RxNav.getAllConcepts — Method

getAllConcepts

/allconcepts All RxTerms concepts

RxNav.getAllConceptsByStatus — Function

getAllConceptsByStatus

/allstatus Concepts having a specified status Current and Historical

RxNav.getAllConceptsByTTY — Method

getAllConceptsByTTY

/allconcepts Concepts having a specified term type Active

RxNav.getAllHistoricalNDCs — Method

getAllHistoricalNDCs

/rxcui/rxcui/allhistoricalndcs National Drug Codes (NDC) ever associated with a concept Current and Historical

RxNav.getAllNDCsByStatus — Function

getAllNDCsByStatus

/allNDCstatus NDCs having a specified NDC status Current and Historical

RxNav.getAllProperties — Function

getAllProperties

/rxcui/rxcui/allProperties Concept details Active

RxNav.getAllRelatedInfo — Method

getAllRelatedInfo

/rxcui/rxcui/allrelated Concepts related directly or indirectly to a specified concept Active

RxNav.getAllRxTermInfo — Method

getAllRxTermInfo

/rxcui/rxcui/allinfo RxTerms information for a specified RxNorm concept Returns the (quite variable) properties data in XML form.

RxNav.getApproximateMatch — Function

getApproximateMatch

/approximateTerm Concept and atom IDs approximately matching a query Active or Current

RxNav.getClassByRxNormDrugId — Function

getClassByRxNormDrugId

/class/byRxcui Classes containing a specified drug RXCUI

RxNav.getClassByRxNormDrugName — Function

getClassByRxNormDrugName

/class/byDrugName Classes containing a drug of the specified name

RxNav.getClassContexts — Method

getClassContexts

/classContext Paths from the specified class to the root of its class hierarchies

RxNav.getClassGraphBySource — Function

getClassGraphBySource

/classGraph Classes along the path from a specified class to the root of a class hierarchy

RxNav.getClassMembers — Function

getClassMembers

/classMembers Drug members of a specified class

RxNav.getClassTree — Function

getClassTree

/classTree Subclasses or descendants of the specified class

RxNav.getClassTypes — Method

getClassTypes

/classTypes Class types

RxNav.getDisplayTerms — Method

getDisplayTerms

/displaynames Strings to support auto-completion in a user interface Active

RxNav.getDrugs — Method

getDrugs

/drugs Drugs related to a specified name Active

RxNav.getIdTypes — Method

getIdTypes

/idtypes Identifier types Current

RxNav.getMultiIngredBrand — Method

getMultiIngredBrand

/brands Brands containing specified ingredients Active

RxNav.getNDCProperties — Method

getNDCProperties

/ndcproperties National Drug Code (NDC) details Active

RxNav.getNDCStatus — Function

getNDCStatus

/ndcstatus Status of a National Drug Code (NDC) Current and Historical

RxNav.getNDCs — Method

getNDCs

/rxcui/rxcui/ndcs National Drug Codes (NDC) associated with a concept Active

RxNav.getPropCategories — Method

getPropCategories

/propCategories RxNav property categories Active

RxNav.getPropNames — Method

getPropNames

/propnames Property names Active

RxNav.getProprietaryInformation — Function

getProprietaryInformation

/rxcui/rxcui/proprietary Strings from sources that require a UMLS license Current

RxNav.getRelaTypes — Method

getRelaTypes

/relatypes RxNorm Relationship types Active

RxNav.getRelas — Method

getRelas

/relas Relationships expressed by a source of drug relations

RxNav.getRelatedByRelationship — Method

getRelatedByRelationship

/rxcui/rxcui/related?rela=... Concepts directly related to a specified concept by a specified relationship Active

RxNav.getRelatedByType — Method

getRelatedByType

/rxcui/rxcui/related?tty=... Concepts of specified types that are directly or indirectly related to a specified concept Active

RxNav.getRxConceptProperties — Method

getRxConceptProperties

/rxcui/rxcui/properties Concept name, TTY, and a synonym Active

RxNav.getRxNormName — Method

getRxNormName

/rxcui/rxcui Name of a concept Active

RxNav.getRxNormVersion — Method

getRxNormVersion

/version RxNorm data set and API versions Current

RxNav.getRxProperty — Method

getRxProperty

/rxcui/rxcui/property A property of a concept Active

RxNav.getRxTermDisplayName — Method

getRxTermDisplayName

/rxcui/rxcui/name RxTerms display name for a specified RxNorm concept

RxNav.getRxTermsVersion — Method

getRxTermsVersion

/version RxTerms version

RxNav.getRxcuiHistoryStatus — Method

getRxcuiHistoryStatus

/rxcui/rxcui/historystatus Status, history, and other attributes of a concept Current and Historical Returns the (quite variable) metadata in XML form.

RxNav.getSimilarityInformation — Function

getSimilarityInformation

/class/similarInfo Similarity of the clinically-significant membership of two classes

RxNav.getSourceTypes — Method

getSourceTypes

/sourcetypes Vocabulary sources Current

RxNav.getSourcesOfDrugClassRelations — Method

getSourcesOfDrugClassRelations

/relaSources Sources of drug-class relations

RxNav.getSpellingSuggestions — Function

getSpellingSuggestions

/spellingsuggestions Drug or class names similar to a given string

RxNav.getTermTypes — Method

getTermTypes

/termtypes Term types Active

RxNav.getdoc — Method

getdoc(urlkey, arg)

get the XML document found by the string formed by: RESTuri[urlkey] * (the urltail)

RxNav.getjson — Method

getjson(urlkey, urltail, APIKEY = "")

Get the JSON document found by the string formed by: uri, arg. Optionally add headers as a vector of Pairs to the request

RxNav.interact — Method

interact(list::Vector)
interact(s1::String, severeonly::Bool=true)
interact(s1::String, s2::String, args...)

Get a list of interactions for a single drug (or rxcui drug id) or pairwise interactions for more than one drug (or rxcuid). Since RxNav no longer maintains an interactions database, this function uses the RxCheck API, and requires an API key to be set in the environment as ENV["RXCHECKAPIKEY"]. See https://rxcheck.dev for more information.

RxNav.interaction — Function

interaction(id, severeonly=true)

Given a drug name or rxcui id string, return known drug interactions for that drug.

NOTE: Because the RxNav database dropped its interactions database in 2023, the RxCheck database is queried here instead. This requires the RxCheck API key be placed in the environment prior to running this function, for example as ENV["RXCHECKAPIKEY"]. A limited-use (100x/month) API key can be obtained from RxCheck.dev for free use as of 2026.

Returns a Vector of NamedTuples as in (drug1, drug2, severity, description)

RxNav.interaction_within_list — Method

interaction_within_list(namelist::Vector{String})

Given a list of drug names or rxcui id strings, return known drug interations for that combination of drugs. Results are organized pairwise, so if A, B, and C have mutual interactions this will be reported for example as A with B, A with C, B with C.

NOTE: Because the RxNav database dropped the interactions database in 2023, the RxCheck database is queried here instead. This requires the RxCheck API key be placed in the environment prior to running this function, for example as ENV["RXCHECKAPIKEY"]. A limited-use (100x/month) API key can be obtained from RxCheck.dev for free use as of 2026.

Returns a Vector of NamedTuples as in (drug1, drug2, severity, description)

RxNav.interaction_within_pair — Method

interaction_within_pair(name1, name2)

Given two drug names or rxcui id strings, return known drug interactions for that pair. Returns a Vector of NamedTuples as in (drug1, drug2, severity, description)

RxNav.is_in_rxcui_format — Method

is_in_rxcui_format(s)

Tests whether the string is in the format for an RxNav rxcui identifier. Currently rxcui identifiers are composed of only digits 0 through 9, though there is nothing in the schema that says these must be only digits. If that changes in future RxNav updates, the parsing here may also change.

RxNav.loadAPIkey — Function

loadAPIkey(printwarning = true)
    Load the RxCheck API key from the environment variable "RXCHECK_API_KEY".
    Returns the API key as a string, or an empty string if not found.

RxNav.morearg — Method

morearg(name, value)

Return further arguments to the REST query of the form "&name=value" or if values is a vector, of the form "&name=val1+val2+val3" Starts with &, not ?, so there must be a prior argument in the url string.

RxNav.morearg — Method

morearg(pairs)

Return further arguments to the REST query of the form "&name=value&name2=value2", etc, where pairs is a vector of [name, value] pairs. Starts with &, not ?, so there must be a prior argument in the url string.

RxNav.name — Method

name(id)

Given the rxcui number as an integer or string, return the drug name, or nothing on failure.

RxNav.rxcui — Method

rxcui(name)

Take a name of an NDC drug, return its rxcui as String, or nothing on failure.